# docs/endpoints/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ithina Admin Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/roles": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "List roles, pre-grouped by audience",
        "description": "Returns the role catalogue pre-grouped into platform_roles and tenant_roles blocks (deliberate D-30 exception). PLATFORM JWTs see both blocks populated; TENANT JWTs see platform_roles always {items: [], total: 0} (audience filter applied at the app layer — these tables have no RLS). Each item carries a user_count field counting active assignments referencing the role (RLS-scoped for TENANT JWTs to the calling tenant).",
        "operationId": "list_roles_api_v1_roles_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: ACTIVE (default), INACTIVE, or ARCHIVED.",
              "title": "Status"
            },
            "description": "Filter by status: ACTIVE (default), INACTIVE, or ARCHIVED."
          },
          {
            "name": "is_system",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Ithina-system flag (true=system roles like SUPER_ADMIN that should not be deleted).",
              "title": "Is System"
            },
            "description": "Filter by Ithina-system flag (true=system roles like SUPER_ADMIN that should not be deleted)."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring match across name, code, and description.",
              "title": "Q"
            },
            "description": "Case-insensitive substring match across name, code, and description."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort key: name_asc (default), name_desc, created_at_asc, created_at_desc. Sort applies within each block.",
              "default": "name_asc",
              "title": "Sort"
            },
            "description": "Sort key: name_asc (default), name_desc, created_at_asc, created_at_desc. Sort applies within each block."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Pagination limit. Present for consistency; v0 catalogue fits in one page.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Pagination limit. Present for consistency; v0 catalogue fits in one page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roles/{role_id}/permissions": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "List permissions granted by a role",
        "description": "Returns the role's granted permissions plus a parent-echo envelope (role_id + role_name at top level, items array below). No pagination — a role has bounded permissions (typically 5-30). TENANT JWTs requesting a PLATFORM-audience role's id receive 404 ROLE_NOT_FOUND (audience filter applied at the app layer; same anti-information-disclosure intent as RLS-as-404 per D-17).",
        "operationId": "list_role_permissions_api_v1_roles__role_id__permissions_get",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermissionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roles/{role_id}": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "Get role detail with held + available permissions",
        "description": "Returns role metadata plus held permissions and grantable permissions (catalogue minus held) for the role-edit screen. Both permission lists carry display labels resolved server-side via JOIN against the lookups table. TENANT-audience roles see ``available_permissions`` with ``scope='GLOBAL'`` rows excluded (audience-scope coherence per LD2). TENANT JWTs requesting a PLATFORM-audience role's id receive 404 ROLE_NOT_FOUND (audience filter applied at the app layer; same anti-information-disclosure intent as RLS-as-404 per D-17). No pagination — the v0 catalogue (~36 permissions) fits in one response.",
        "operationId": "get_role_api_v1_roles__role_id__get",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "rbac"
        ],
        "summary": "Edit a role's name, description, and/or permission set",
        "description": "Partial update of a role. Editable fields: name, description, permission_ids (replace-set; diff-replace preserves created_at on unchanged role_permissions rows per LD5). audience, code, is_system, status, and audit columns are rejected at the schema layer via extra='forbid'. Gated by ADMIN.ROLES.OVERRIDE.GLOBAL (held by SUPER_ADMIN only post Step 6.18.1 seed; PLATFORM-only by gate-tuple construction). SUPER_ADMIN role itself is uneditable via API in v0 (409 SUPER_ADMIN_PROTECTED); operator workflow is direct SQL on core.roles / core.role_permissions. Two-layer OVERRIDE.GLOBAL invariant guards against zeroing out the platform-admin bootstrap.",
        "operationId": "patch_role_api_v1_roles__role_id__patch",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Role Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permissions": {
      "get": {
        "tags": [
          "rbac"
        ],
        "summary": "List permission catalogue",
        "description": "Returns the canonical permission catalogue ((module, resource, action, scope) tuples). Reference data — both user types see all rows; no audience filter. Default sort clusters related permissions together (module_asc compound: module/resource/action/scope ASC), matching matrix render order.",
        "operationId": "list_permissions_api_v1_permissions_get",
        "parameters": [
          {
            "name": "module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by module: ADMIN, PRICING_OS, PERISHABLES_ASSISTANT, PROMOTIONS_ASSISTANT.",
              "title": "Module"
            },
            "description": "Filter by module: ADMIN, PRICING_OS, PERISHABLES_ASSISTANT, PROMOTIONS_ASSISTANT."
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by scope: GLOBAL, TENANT, STORE.",
              "title": "Scope"
            },
            "description": "Filter by scope: GLOBAL, TENANT, STORE."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort key: module_asc (default; compound module/resource/action/scope ASC), code_asc, code_desc.",
              "default": "module_asc",
              "title": "Sort"
            },
            "description": "Sort key: module_asc (default; compound module/resource/action/scope ASC), code_asc, code_desc."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Pagination limit. Present for consistency; v0 catalogue fits in one page.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Pagination limit. Present for consistency; v0 catalogue fits in one page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleListResponse": {
        "properties": {
          "platform_roles": {
            "$ref": "#/components/schemas/AudienceBlock",
            "description": "PLATFORM-audience roles (Ithina staff). Empty for TENANT JWTs."
          },
          "tenant_roles": {
            "$ref": "#/components/schemas/AudienceBlock",
            "description": "TENANT-audience roles (customer staff)."
          }
        },
        "type": "object",
        "required": [
          "platform_roles",
          "tenant_roles"
        ],
        "title": "RoleListResponse",
        "description": "E1 response: roles pre-grouped by audience (deliberate D-30 exception).\n\nFor TENANT JWTs the ``platform_roles`` block always returns\n``{items: [], total: 0}``. The frontend suppresses the empty\nsection header."
      },
      "RolePermissionsResponse": {
        "properties": {
          "role_id": {
            "type": "string",
            "format": "uuid",
            "title": "Role Id",
            "description": "Echo of the path parameter. Frontend race-condition guard: if the user clicks two roles in quick succession, the older response can be discarded by id-equality without a round-trip."
          },
          "role_name": {
            "type": "string",
            "title": "Role Name",
            "description": "Display name of the role. Saves the frontend a cross-lookup against E1's cached list when rendering the right-pane header."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PermissionRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "role_id",
          "role_name",
          "items"
        ],
        "title": "RolePermissionsResponse",
        "description": "E3 response: parent role identity echoed + permissions list.\n\nNo pagination — a role has bounded permissions (typically 5-30);\nthe parent echo (``role_id`` + ``role_name``) lets the frontend\navoid an additional E1 lookup when rendering the right-pane header."
      },
...
    }
  }
}